requirements.txt
django==5.0.2
psycopg2-binary
ruff
pygraphviz
django-extensions
celery
redis
djangorestframework
requests
django-celery-beat
requests
